+++ domain/src/main/java/com/example/domain/interactor/order/ICreateOrderInteractor.java
@@ -3,6 +3,8 @@ package com.example.domain.interactor.order;
 import com.example.domain.model.OrderItem;
 import com.example.domain.model.Product;
 
+import java.math.BigDecimal;
+
 import io.reactivex.Observable;
 
 public interface ICreateOrderInteractor {
@@ -10,5 +12,6 @@ public interface ICreateOrderInteractor {
 
     void finish();
 
-    void orderItemAdded(Product product, float selectedWeight, double sum);
+    void orderItemAdded(Product product, float selectedWeight, BigDecimal sum);
+
 }
